refactoring

+++ Shooter/Assets/Scripts/HitBox/UnitHitBox.cs
@@ -6,19 +6,19 @@ public class UnitHitBox : HitBox
 
     public override void Visit(RifleAttack attack, RaycastHit hit)
     {
-        ApplyDamage(attack.Damage, 1);
+        ApplyDamage(attack.Damage, 1f);
 
         print("You are got on unit from shotgun");
     }
 
     public override void Visit(ShotgunAttack attack, RaycastHit hit)
     {
-        ApplyDamage(attack.Damage, 1);
+        ApplyDamage(attack.Damage, 1f);
 
         print("You are got on unit from shotgun");
     }
 
-    private void ApplyDamage(int damage, int multiplayer)
+    private void ApplyDamage(float damage, float multiplayer)
     {
         _health.TakeHealth(damage * multiplayer);
     }
